+++ buildpacks/maven/src/layer/maven.rs
@@ -1,5 +1,5 @@
 use crate::util::extract_tarball;
-use crate::{MavenBuildpack, MavenBuildpackError, Tarball};
+use crate::{MavenBuildpack, MavenBuildpackError, MavenVersion};
 use libcnb::build::BuildContext;
 use libcnb::data::layer_content_metadata::LayerTypes;
 use libcnb::layer::{ExistingLayerStrategy, Layer, LayerData, LayerResult, LayerResultBuilder};
@@ -8,14 +8,16 @@ use libcnb::Buildpack;
 use serde::{Deserialize, Serialize};
 use std::fs::File;
 use std::path::Path;
+use url::Url;
 
 pub(crate) struct MavenLayer {
-    pub(crate) tarball: Tarball,
+    pub(crate) apache_maven_mirror: Url,
+    pub(crate) version: MavenVersion,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub(crate) struct MavenLayerMetadata {
-    tarball: Tarball,
+    version: MavenVersion,
 }
 
 impl Layer for MavenLayer {
@@ -38,17 +40,19 @@ impl Layer for MavenLayer {
         let temp_dir = tempfile::tempdir().unwrap();
         let temp_file_path = temp_dir.path().join("maven.tar.gz");
 
-        libherokubuildpack::download::download_file(&self.tarball.url, &temp_file_path)
+        let maven_url = self.apache_maven_mirror.join(&self.version.path).unwrap();
+
+        libherokubuildpack::download::download_file(maven_url, &temp_file_path)
             .map_err(MavenBuildpackError::MavenTarballDownloadError)?;
 
         libherokubuildpack::digest::sha256(&temp_file_path)
             .map_err(MavenBuildpackError::MavenTarballSha256IoError)
             .and_then(|downloaded_tarball_sha256| {
-                if downloaded_tarball_sha256 == self.tarball.sha256 {
+                if downloaded_tarball_sha256 == self.version.sha256 {
                     Ok(())
                 } else {
                     Err(MavenBuildpackError::MavenTarballSha256Mismatch {
-                        expected_sha256: self.tarball.sha256.clone(),
+                        expected_sha256: self.version.sha256.clone(),
                         actual_sha256: downloaded_tarball_sha256,
                     })
                 }
@@ -69,7 +73,7 @@ impl Layer for MavenLayer {
         );
 
         LayerResultBuilder::new(MavenLayerMetadata {
-            tarball: self.tarball.clone(),
+            version: self.version.clone(),
         })
         .env(layer_env)
         .build()
@@ -80,7 +84,7 @@ impl Layer for MavenLayer {
         _context: &BuildContext<Self::Buildpack>,
         layer_data: &LayerData<Self::Metadata>,
     ) -> Result<ExistingLayerStrategy, <Self::Buildpack as Buildpack>::Error> {
-        let strategy = if layer_data.content_metadata.metadata.tarball == self.tarball {
+        let strategy = if layer_data.content_metadata.metadata.version == self.version {
             ExistingLayerStrategy::Keep
         } else {
             ExistingLayerStrategy::Recreate